Use "cast" filters for numeric variables used in Jinja numeric operations
Variables set by users in the public API are not guaranteed to be any
specific data type, and may be `str` type when some numeric type is
expected.  Use `int`, `float`, and `bool` filters to guarantee that
the variable is the correct data type.
Instead of `int_var < 1`, use `int_var | int < 1`.